updated  get_uses_emails router

+++ app/utils.py
@@ -24,7 +24,7 @@ def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
         expire = datetime.utcnow() + expires_delta
     else:
         expire = datetime.utcnow() + timedelta(
-            hours=float(os.getenv("ACCESS_TOKEN_EXPIRATION_TIME"))
+            hours=int(os.getenv("ACCESS_TOKEN_EXPIRATION_TIME"))
         )
     to_encode.update({"exp": expire})
     encoded_jwt = jwt.encode(
